Rename malegislature package to ma_legislature
